requirements.txt
aiohttp==3.9.1
aiohttp-apispec==2.2.3
aiohttp-autoreload==0.0.1
aiohttp-middlewares==2.2.1
aiosignal==1.3.1
apispec==3.3.2
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
black==23.12.1
click==8.1.7
flake8==7.0.0
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
isort==5.13.2
Jinja2==3.1.3
MarkupSafe==2.1.4
marshmallow==3.20.2
marshmallow-sqlalchemy==0.30.0
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pycodestyle==2.11.1
pyflakes==3.2.0
python-dotenv==1.0.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
webargs==5.5.3
yarl==1.9.4
